requirements.txt
streamlit
pandas
numpy
scipy
python-dotenv
langchain
langchain-groq
openai
streamlit-lottie
